fixed add product with image

+++ src/middlewares/cloud_file_handler.js
@@ -1,4 +1,5 @@
 const fs = require('fs');
+const path = require('path');
 
 const imageCompressor = require('../utils/image_compressor');
 const cloudFileUploader = require('../utils/cloud_file_util');
@@ -7,25 +8,32 @@ const cloudFileHandler = async (req, res, next) => {
   try {
     if (req.files) {
       const imageUrlList = [];
+
       for (index in req.files) {
         const compressedFile = await imageCompressor.compressImage(
-          req.files[index].path
+          req.files[index]
         );
-        const filePath = compressedFile.path;
+
+        if (!compressedFile) {
+          return next();
+        }
+
+        const filePath = path.join(
+          __dirname + '../../../data/uploads',
+          req.files[index].originalname
+        );
+
         const { url } = await cloudFileUploader(filePath, 'productImages');
         fs.unlinkSync(filePath);
+
         if (url) {
           imageUrlList.push({ url });
         }
       }
       req.imageUrls = imageUrlList;
-      next();
-    } else {
-      next();
     }
-  } catch (error) {
-    next();
-  }
+  } catch (error) {}
+  next();
 };
 
 module.exports = cloudFileHandler;
